Add syntactic sugar for framing.
